requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
google-generativeai>=0.3.0
pillow>=10.0.0